requirements.txt
openai
langchain
streamlit
python-dotenv
google-generativeai

